Move adapter class to separate module and update imports
