+++ sailon/__init__.py
@@ -1,3 +1,4 @@
 from .ond_adapter import ONDAdapter
 from .random_novelty_detector import RandomNoveltyDetector
+from .random_novelty_detector_adapter import RandomNoveltyDetectorAdapter
 from .dummy_interface import DummyInterface
